pyproject.toml
[project]
name = "archetype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "getdaft>=0.4.13",
    "pyarrow>=16.0.0",
    "ray[default] >= 2.20.0,<2.31.0",
    "pyiceberg[glue] == 0.9.0",
    "lancedb>=0.22.0",
    "python-ulid>=3.0.0",
    "pytest>=8.3.5",
]
